+++ app/dashboard/vite.config.ts
@@ -1,9 +1,11 @@
 import react from "@vitejs/plugin-react";
-import { copyFileSync } from "node:fs";
+import { copyFileSync, readFileSync } from "node:fs";
 import { defineConfig, splitVendorChunkPlugin } from "vite";
 import svgr from "vite-plugin-svgr";
 import { visualizer } from "rollup-plugin-visualizer";
 import tsconfigPaths from "vite-tsconfig-paths";
+
+const buildVersion = readFileSync("../../VERSION", "utf8").trim();
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -13,7 +15,7 @@ export default defineConfig({
     outDir: "build",
   },
   define: {
-    __LOCALE_BUILD_ID__: JSON.stringify(Date.now().toString()),
+    __LOCALE_BUILD_ID__: JSON.stringify(buildVersion),
   },
   plugins: [
     tsconfigPaths(),